unittest: Fix compiler warnings (signed/unsigned)

+++ unittest/baseapi_test.cc
@@ -319,7 +319,7 @@ TEST_F(TesseractTest, InitConfigOnlyTest) {
   const char* langs[] = {"eng", "chi_tra", "jpn", "vie"};
   std::unique_ptr<tesseract::TessBaseAPI> api;
   CycleTimer timer;
-  for (int i = 0; i < ARRAYSIZE(langs); ++i) {
+  for (size_t i = 0; i < ARRAYSIZE(langs); ++i) {
     api.reset(new tesseract::TessBaseAPI);
     timer.Restart();
     EXPECT_EQ(0, api->Init(TessdataPath().c_str(), langs[i],
@@ -333,7 +333,7 @@ TEST_F(TesseractTest, InitConfigOnlyTest) {
   vars_vec.push_back(STRING("tessedit_init_config_only"));
   vars_values.push_back(STRING("1"));
   LOG(INFO) << "Switching to config only initialization:";
-  for (int i = 0; i < ARRAYSIZE(langs); ++i) {
+  for (size_t i = 0; i < ARRAYSIZE(langs); ++i) {
     api.reset(new tesseract::TessBaseAPI);
     timer.Restart();
     EXPECT_EQ(0, api->Init(TessdataPath().c_str(), langs[i],
